docs/requirements.txt
# Project 1: Requirements and Dependencies
# File: requirements.txt
# By: Islam Zain

# Core AWS SDK for Python
boto3==1.28.85
botocore==1.31.85

# Image Processing Library
Pillow==10.0.0

# Optional dependencies for enhanced functionality
# Uncomment as needed

# For advanced image processing
# opencv-python==4.8.0.76
# scikit-image==0.21.0

# For metadata extraction
# Pillow[webp]==10.0.0
# piexif==1.1.3

# For logging and monitoring
# python-json-logger==2.0.7
# aws-xray-sdk==2.12.0

# Development dependencies (for testing locally)
# pytest==7.4.2
# moto[s3,sns]==4.2.9
# black==23.10.1
# pylint==3.0.0
